BASIC ALTERNATE ENGINE SUPPORT
Basic support fo alternate attack engines has been added. By creating
a file called <enginename>_engine.py in the beeswithmachineguns
directory (with bees.py and main.py), you can utilize it to provide
the attack/results data while leveraging the rest of the existing
bees code.

To help facilitate this, a new parameter was added to the params
passed to _attack that provides the keypath.